backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
boto3==1.34.129
requests-oauthlib==2.0.0
cryptography==46.0.7
python-dotenv==1.0.1
pymongo==4.6.3
pydantic==2.6.4
email-validator==2.2.0
pyjwt==2.12.0
bcrypt==4.1.3
passlib==1.7.4
tzdata==2024.2
motor==3.3.1
python-jose==3.4.0
requests==2.33.0
pandas==2.2.0
python-multipart==0.0.22
typer==0.9.0
arq==0.25.0
redis==5.0.0
slowapi==0.1.9
httpx==0.27.0
sentry-sdk[fastapi]==2.8.0
google-auth==2.29.0
aiofiles==24.1.0
aiosmtplib==3.0.2
openpyxl==3.1.5
reportlab==4.2.5
